Move dtype conversion to pandas_type_compatibility
